feat: rebuild WheelAthlete 1.8.4 candidate

+++ applications/wheelathlete_windows/tools/pc_gui/process_manager.py
@@ -98,6 +98,8 @@ class DaemonProcessManager(QObject):
             return
         if self.process.state() == QProcess.ProcessState.NotRunning:
             return
+        if self.process.waitForFinished(2500):
+            return
         self.process.terminate()
         if not self.process.waitForFinished(2000):
             self.process.kill()
